ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.4.0
numpy==1.26.3
nltk==3.8.1
python-multipart==0.0.6
pydantic==2.5.3
requests==2.31.0
openai==1.12.0
google-generativeai==0.4.0
pytest==8.0.0
httpx==0.27.0
